Safari-609.1.11.1/LayoutTests/fast/events/focus-label-legend-elements-with-tab-expected.txt

Tests for moving focus by pressing tab key across label and legend elements.
To manually test, press tab key seven times then shift+tab six times.
It should traverse focusable elements in the ascending order from 1 to 7 and then in the descending order back to 1.


1. An input element inside a non-focusable label element
2. A label element with an input element
3. An input elment inside a focusable label element
4. A label element with just text
6. A focusable legend element inside a fieldset element
7. An input element inside a fieldset element
7. An input element inside a fieldset element
6. A focusable legend element inside a fieldset element
4. A label element with just text
3. An input elment inside a focusable label element
2. A label element with an input element
1. An input element inside a non-focusable label element
